Create a separate types files to store common types
