+++ lte/gateway/c/session_manager/CMakeLists.txt
@@ -172,6 +172,7 @@ add_library(SESSION_MANAGER
     Utilities.h
     OperationalStatesHandler.cpp
     OperationalStatesHandler.h
+    Types.h
     ${PROTO_SRCS}
     ${PROTO_HDRS})
